requirements.txt
appdirs==1.4.4
argcomplete==3.1.2
blurhash==1.1.4
certifi==2023.7.22
charset-normalizer==3.3.1
decorator==5.1.1
idna==3.4
Mastodon.py==1.8.1
Pillow==10.1.0
pypng==0.20220715.0
pyserial==3.5
python-dateutil==2.8.2
python-escpos==2.2.0
python-magic==0.4.27
pyusb==1.2.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
